+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.9.5]
 * Fixed compatibility with Emscripten 3.1.44
 * Improved support for `Pointer<Utf8>`
+* Added missing types `UintPtr`, `Bool`, `Int` and `UnsignedInt`
 
 ## [0.9.4]
 * Fixed analyzer warnings